requirements.txt
Django==4.2
djangorestframework==3.14.0
mysqlclient==2.1.1
gunicorn==20.1.0
